+++ test/perm_grp_algebra.jl
@@ -88,5 +88,4 @@
         end
         @test a ≤ b
     end
-
 end
